requirements_dev.txt
Django==4.0.5
django-bootstrap5==21.3
Pygments==2.12.0
django-pygments-renderer==0.0.1
django-extensions==3.1.5
pytest-bigquery-mock
pandas
